backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
aiofiles==24.1.0

# LangChain & LLM
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
openai==1.57.4

# Vector DB & Embeddings
chromadb==0.5.23
sentence-transformers==3.3.1

# PDF Processing
pymupdf==1.24.14

# Content Generation
python-pptx==1.0.2
google-api-python-client==2.154.0

# HTTP & Utils
httpx==0.28.1
aiosqlite==0.20.0
